# Export cache image and app image in parallel