In our scenario, the app image and the cache image need to be exported at the same time, but this process is serial in the lifecycle, which means that after the app image is exported, we have to wait for the cache image to be exported. After calculation, the time to export the app image is about the same as the time to export the cache, but we don’t need to wait for the export of cache image, only after the app is exported, we can proceed to the next steps (distribution and deployment).

So we tried to parallelize this step (this PR) and compare it with the serial exporting. We used several projects for testing and pushed app images and cache images to the same self-hosting registry.

- Java (app image is 202.361MB, cache image is 157.525MB, with a same layer: 107.648MB): 
  - Before: total 18.34s, app 8.96s, cache 9.38s
  - After: total 14.70s, app 11.42s, cache 13.93s
  - app image: 0+1.103MB+15.153MB+107.648MB+49.953MB+0+0+28.502MB
  - cache image: 9.411MB+40.465MB+107.648MB

- Go (app image is 114.273MB, cache is 175.833MB, no same layer): 
  - Before: total 16.57s, app 5.92s, cache 10.65s
  - After: total 12.02s, app 7.31s, cache 11.48s
  - app image: 0+1MB+25.72MB+8.993MB+49.953MB+0+0+28.502MB
  - cache image: 70.87MB+104.964MB

We can get some improvements here, although the export time of app image and cache image has significantly increased, the total time has decreased. The effect will be more pronounced if we export to different registry or use faster bandwidth.

But my confusion is, if the app image and cache image contain same layeres, will this method no longer resue layeres when pushing to the registry. Or we should detece / specify whether to use a parallel pushing strategy. Thx.