@joe-kimmel-vmw 2023-07-31
LGTM;  I'm +1 on using `defer` as Jesse suggested.