@ESWZY 2023-08-01
> Thanks for your contribution @ESWZY ! Would be awesome if you could add some unit tests to validate that your changes work as expected. You can find utilities to test async code in the `testhelpers.go` file, specifically the helper function `Eventually`.

Thanks for your comment! It is reasonable to add tests for the parallelized export process. But I found that most of the existing tests are to test the two steps of `func (e *Exporter) Export` and `func (e *Exporter) Cache`; or to run a Docker container to test the entire export process. 

But the modified parallel logic is in between, should I reuse the existing testing logic? I found some test cases here that I feel are similar, but not sure if I should create a new test case based on this, or just reuse this case and modify it:

https://github.com/buildpacks/lifecycle/blob/main/acceptance/exporter_test.go#L294-318

```go
// Add these lines to detect whether the export of the two is successful
h.Run(t, exec.Command("docker", "pull", cacheImageName))
h.Run(t, exec.Command("docker", "pull", exportedImageName))
```