@ESWZY 2023-08-14
> @ESWZY is this ready for a re-review?

@natalieparellano @dlion  Sorry for no reply for a long time. I was trying to design a test case for the parallel export process, but had no idea how to do that. 

The original test cases have already included the export tests of single app image, single cache image and both of them. These test cases have been able to part of prove that the parallel export process works normally. Should we add some more complex, or more intensive export tests? Or just use the original test cases without any modification. Thx.