@natalieparellano 2023-08-16
I've been thinking about this more and I think we need to check that `launch = true` `cache = true` layers are cached successfully when this is done in parallel. I fear the cacher may be relying on the app exporter for creating these layers.