@natalieparellano 2023-08-17
Looking into the code a little bit I think we'll want to put a mutex around `e.LayerFactory.DirLayer` so that we avoid re-creating the same layer.tar in parallel in the case that the exporter (while exporting) and the exporter (while caching) is processing the same layer.