@natalieparellano 2023-08-17
> we'll want to put a mutex around `e.LayerFactory.DirLayer`

I think we can make `Factory.tarHashes` a `sync.Map` and use `LoadOrStore` when we access the map here: https://github.com/buildpacks/lifecycle/blob/df9ae90e17bc96aeac492634ce551d7d4e68755a/layers/factory.go#L45