@natalieparellano 2023-08-21
> Maybe something like EXPORTED_APP_IMAGE_REF that contains the registry.blah/repo/blah@digest would useful

Another option would be to look for the presence of <layers>/report.toml, which is currently [written](https://github.com/buildpacks/lifecycle/blob/c72dc20a932a7e9a2fb300798ef07c99feac8178/cmd/lifecycle/exporter.go#L224-L226) when the app image is done, but before we start processing the cache image