@ESWZY 2023-08-23
> @ESWZY we could add that! It would require an RFC: https://github.com/buildpacks/rfcs#rfc-process
> 
> Is that something you'd be willing to contribute? We could guide you through the process if that would be helpful.

I would like to! Let me read [RFC 0004](https://github.com/buildpacks/rfcs/blob/main/text/0004-rfc-process.md) first. 🥰